Consolidate FeatureFlag functions into BySite and ByUser
