+++ src/api/Nhs.Appointments.Api/Features/Flags.cs
@@ -2,9 +2,16 @@ namespace Nhs.Appointments.Api.Features;
 
 public static class Flags
 {
+    #region TestFlags
+    //a simple on/off global flag
     public const string TestFeatureEnabled = "TestFeatureEnabled";
+    //a flag configured to be enabled for a pre-defined time period
     public const string TestFeatureTimeWindowEnabled = "TestFeatureTimeWindowEnabled";
+    //a flag configured to return enabled/disabled randomly on each request at the provided probability percentage
     public const string TestFeaturePercentageEnabled = "TestFeaturePercentageEnabled";
+    //a flag configured to be enabled for a specific cohort of users
     public const string TestFeatureUsersEnabled = "TestFeatureUsersEnabled";
+    //a flag configured to be enabled for a specific collection of sites
     public const string TestFeatureSitesEnabled = "TestFeatureSitesEnabled";
+    #endregion
 }
