+++ src/api/Nhs.Appointments.Api/Functions/GetFeatureFlagsBySiteFunction.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using System.Net;
+using System.Reflection;
 using System.Threading.Tasks;
 using FluentValidation;
 using Microsoft.AspNetCore.Http;
@@ -8,6 +9,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.Azure.WebJobs.Extensions.OpenApi.Core.Attributes;
 using Microsoft.Extensions.Logging;
+using Microsoft.FeatureManagement;
 using Nhs.Appointments.Api.Features;
 using Nhs.Appointments.Api.Models;
 using Nhs.Appointments.Core;
@@ -15,37 +17,50 @@ using Nhs.Appointments.Core.Inspectors;
 
 namespace Nhs.Appointments.Api.Functions;
 
-public class GetFeatureFlagPercentageFunction(
+public class GetFeatureFlagsBySiteFunction(
     IValidator<EmptyRequest> validator,
     IUserContextProvider userContextProvider,
-    ILogger<GetFeatureFlagPercentageFunction> logger,
+    ILogger<GetFeatureFlagsBySiteFunction> logger,
     IMetricsRecorder metricsRecorder,
     IFeatureToggleHelper featureToggleHelper)
-    : BaseApiFunction<EmptyRequest, bool>(validator, userContextProvider, logger, metricsRecorder)
+    : BaseApiFunction<EmptyRequest, GetFeatureFlagsResponse>(validator, userContextProvider, logger, metricsRecorder)
 {
-    [OpenApiOperation(operationId: "GetFeatureFlagPercentage", tags: ["FeatureFlag"],
-        Summary = "Get the enabled state for the requested flag, using the percentage filter")]
-    [OpenApiResponseWithBody(statusCode: HttpStatusCode.OK, "application/json", typeof(bool),
-        Description = "Information for a single site")]
+    [OpenApiOperation(operationId: "GetFeatureFlagsBySite", tags: ["FeatureFlag"],
+        Summary = "Get the enabled state for all the defined feature flags, using the provided site id for any feature filters")]
+    [OpenApiResponseWithBody(statusCode: HttpStatusCode.OK, "application/json", typeof(GetFeatureFlagsResponse),
+        Description = "The enabled state for all the defined feature flags")]
     [OpenApiResponseWithBody(statusCode: HttpStatusCode.BadRequest, "application/json",
         typeof(IEnumerable<ErrorMessageResponseItem>), Description = "The body of the request is invalid")]
     [OpenApiResponseWithBody(statusCode: HttpStatusCode.Unauthorized, "application/json",
         typeof(ErrorMessageResponseItem), Description = "Unauthorized request to a protected API")]
     [OpenApiResponseWithBody(statusCode: HttpStatusCode.Forbidden, "application/json", typeof(ErrorMessageResponseItem),
         Description = "Request failed due to insufficient permissions")]
-    [Function("GetFeatureFlagPercentageFunction")]
+    [Function("GetFeatureFlagsBySiteFunction")]
     public override Task<IActionResult> RunAsync(
-        [HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "feature-flag/percentage")]
+        [HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "feature-flags/sites/{site}")]
         HttpRequest req, FunctionContext functionContext)
     {
         return base.RunAsync(req, functionContext);
     }
 
-    protected override async Task<ApiResult<bool>> HandleRequest(EmptyRequest request, ILogger logger,
-        FunctionContext functionContext)
+    protected override async Task<ApiResult<GetFeatureFlagsResponse>> HandleRequest(EmptyRequest request,
+        ILogger logger, FunctionContext functionContext)
     {
-        var isFeatureEnabled = await featureToggleHelper.IsFeatureEnabledForFunction(Flags.TestFeaturePercentageEnabled, functionContext, Principal, new NoSiteRequestInspector());
-        return ApiResult<bool>.Success(isFeatureEnabled);
+        var response = new GetFeatureFlagsResponse([]);
+
+        var flags = typeof(Flags)
+            .GetFields(BindingFlags.Public | BindingFlags.Static)
+            .Where(f => f.FieldType == typeof(string))
+            .Select(f => f.GetValue(null)?.ToString())
+            .ToList();
+
+        foreach (var flag in flags)
+        {
+            var enabled = await featureToggleHelper.IsFeatureEnabledForFunction(flag, functionContext, Principal, new SiteFromPathInspector());
+            response.FeatureFlags.Add(new KeyValuePair<string, bool>(flag, enabled));
+        }
+
+        return ApiResult<GetFeatureFlagsResponse>.Success(response);
     }
 
     protected override Task<IEnumerable<ErrorMessageResponseItem>> ValidateRequest(EmptyRequest request)
