+++ src/api/Nhs.Appointments.Api/Functions/GetFeatureFlagsByUserFunction.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using System.Net;
+using System.Reflection;
 using System.Threading.Tasks;
 using FluentValidation;
 using Microsoft.AspNetCore.Http;
@@ -15,38 +16,52 @@ using Nhs.Appointments.Core.Inspectors;
 
 namespace Nhs.Appointments.Api.Functions;
 
-public class GetFeatureFlagByUserFunction(
+public class GetFeatureFlagsByUserFunction(
     IValidator<EmptyRequest> validator,
     IUserContextProvider userContextProvider,
-    ILogger<GetFeatureFlagByUserFunction> logger,
+    ILogger<GetFeatureFlagsByUserFunction> logger,
     IMetricsRecorder metricsRecorder,
     IFeatureToggleHelper featureToggleHelper)
-    : BaseApiFunction<EmptyRequest, bool>(validator, userContextProvider, logger, metricsRecorder)
+    : BaseApiFunction<EmptyRequest, GetFeatureFlagsResponse>(validator, userContextProvider, logger, metricsRecorder)
 {
-    [OpenApiOperation(operationId: "GetFeatureFlagByUser", tags: ["FeatureFlag"],
-        Summary = "Get the enabled state for the requested flag, by user filter")]
-    [OpenApiResponseWithBody(statusCode: HttpStatusCode.OK, "application/json", typeof(bool),
-        Description = "Information for a single site")]
+    [OpenApiOperation(operationId: "GetFeatureFlagsByUser", tags: ["FeatureFlag"],
+        Summary =
+            "Get the enabled state for all the defined feature flags, using the calling users id for any feature filters")]
+    [OpenApiResponseWithBody(statusCode: HttpStatusCode.OK, "application/json", typeof(GetFeatureFlagsResponse),
+        Description = "The enabled state for all the defined feature flags")]
     [OpenApiResponseWithBody(statusCode: HttpStatusCode.BadRequest, "application/json",
         typeof(IEnumerable<ErrorMessageResponseItem>), Description = "The body of the request is invalid")]
     [OpenApiResponseWithBody(statusCode: HttpStatusCode.Unauthorized, "application/json",
         typeof(ErrorMessageResponseItem), Description = "Unauthorized request to a protected API")]
     [OpenApiResponseWithBody(statusCode: HttpStatusCode.Forbidden, "application/json", typeof(ErrorMessageResponseItem),
         Description = "Request failed due to insufficient permissions")]
-    [Function("GetFeatureFlagByUserFunction")]
+    [Function("GetFeatureFlagsByUserFunction")]
     public override Task<IActionResult> RunAsync(
-        [HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "feature-flag/user")]
+        [HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "feature-flags/by-user")]
         HttpRequest req, FunctionContext functionContext)
     {
         return base.RunAsync(req, functionContext);
     }
 
-    protected override async Task<ApiResult<bool>> HandleRequest(EmptyRequest request, ILogger logger,
-        FunctionContext functionContext)
+    protected override async Task<ApiResult<GetFeatureFlagsResponse>> HandleRequest(EmptyRequest request,
+        ILogger logger, FunctionContext functionContext)
     {
-        var isFeatureEnabled = await featureToggleHelper.IsFeatureEnabledForFunction(
-            Flags.TestFeatureUsersEnabled, functionContext, Principal, new NoSiteRequestInspector());
-        return ApiResult<bool>.Success(isFeatureEnabled);
+        var response = new GetFeatureFlagsResponse([]);
+
+        var flags = typeof(Flags)
+            .GetFields(BindingFlags.Public | BindingFlags.Static)
+            .Where(f => f.FieldType == typeof(string))
+            .Select(f => f.GetValue(null)?.ToString())
+            .ToList();
+
+        foreach (var flag in flags)
+        {
+            var enabled = await featureToggleHelper.IsFeatureEnabledForFunction(flag, functionContext, Principal,
+                new NoSiteRequestInspector());
+            response.FeatureFlags.Add(new KeyValuePair<string, bool>(flag, enabled));
+        }
+
+        return ApiResult<GetFeatureFlagsResponse>.Success(response);
     }
 
     protected override Task<IEnumerable<ErrorMessageResponseItem>> ValidateRequest(EmptyRequest request)
